React Routing, new files
I have organized all the components, pages and so on into different folders.

The routing rendering part is now done. I have started with HTML coding for our two different Headers (One header for when you are logged in and one header when you are a "guest" on the website (not signed-in)).